% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_adj.R
\name{c_adj}
\alias{c_adj}
\title{Calculate the means and variance of a normal linear predictor given values of C and prevalence}
\usage{
c_adj(target.prev, target.c, p = NULL, set.seed = 1)
}
\arguments{
\item{target.prev}{(numeric) The target prevalence}

\item{target.c}{(numeric) The target aUC}

\item{check}{(logical) check using a large dataset whether mean and variance correspond to the target values}
}
\value{
adjusted C
}
\description{
This function calculates the adjusted C
}
\examples{

c_adjusted <- c_adj(target.prev=0.1, target.c=0.85, p=10)
c_adjusted
samplesizedev(S=0.9, phi = 0.1, c = c_adjusted, p = 10))

##################################################################
}
