% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_cs.R
\name{expected_cs}
\alias{expected_cs}
\title{Calculate the expected calibration slope (and mean absolute prediction error (MAPE) for binary data) for a given  sample size}
\usage{
expected_cs(
  outcome = "Binary",
  n,
  phi,
  c,
  p,
  gamma = rep(1/p, p),
  nval = 25000,
  nsim = 1000,
  parallel = TRUE,
  method = "MLE"
)
}
\arguments{
\item{outcome}{(character) The type of outcome (''Binary'' or ''Survival'')}

\item{n}{(numeric) The sample size}

\item{c}{(numeric) The anticipated c-statistic (binary outcome) or c-index (survival outcome)}

\item{p}{(numeric) The anticipated outcome prevalence (binary outcome) or proportion of events (survival outcome)}

\item{gamma}{(numeric) Relative strength of predictors (default=rep(1/p,p); same length as p, must sums up to 1)}

\item{nval}{(numeric) Size of validation data (default=25000; use at least 10000)}

\item{nsim}{(numeric) The number of simulations (default=1000; use at least 500 to ensure small simulation error)}

\item{parallel}{(logical) parallel processing to speed up computations (default=TRUE)}

\item{method}{(character) the fitting method (default="MLE"; currently the only option. Other fitting methods will be added in future versions)}

\item{n.predictors}{(numeric) The number of candidate predictor variables}
}
\value{
df: the mean calibration slope (with SD),the probability of calibration slope<0.8,  and mean MAPE (with SD) for binary outcomes.
}
\description{
This function calculates the expected calibration slope and MAPE given key data and model characteristics
(outcome prevalence, c-statistic and number of predictors). It takes approximately 15 seconds to run (binary outcome).
}
\examples{
# expected_cs(outcome="Binary", n = 530, p = 0.2, c = 0.85, n.predictors = 10, nsim = 100, parallel = FALSE)

# Prefer parallel computing with >2 cores that ensure faster running
# expected_cs(n = 530, p = 0.2, c = 0.85, n.predictors = 10, nsim = 100, parallel = TRUE)
}
\seealso{
samplesizedev
}
