% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_cs_survival.R
\name{expected_cs_survival}
\alias{expected_cs_survival}
\title{Calculate the expected calibration slope and mean for a given  sample size}
\usage{
expected_cs_survival(
  n,
  p,
  c,
  n.predictors,
  beta = rep(1/n.predictors, n.predictors),
  nsim = 1000,
  nval = 25000,
  parallel = TRUE,
  method = "MLE"
)
}
\arguments{
\item{n}{(numeric) The sample size}

\item{p}{(numeric) The anticipated proportion of events at the time-point of interest}

\item{c}{(numeric) The C-index}

\item{n.predictors}{(numeric) The number of candidate predictor variables}

\item{nsim}{(numeric) The number of simulations (at least 500, default value 1000)}

\item{nval}{(numeric) Size of validation data}

\item{parallel}{(logical) parallel processing to speed up computations (default=TRUE)}
}
\value{
df: the expected calibration slope
}
\description{
This function calculates the expected calibration slope given key data and model characteristics
(proportion of censoring at time point of interest, C-index and number of predictors).
}
\examples{
# expected_cs_survival(n = 530, p = 0.2, c = 0.85, n.predictors = 10, nsim = 100, parallel = FALSE)

# Prefer parallel computing with >2 cores that ensure faster running
# expected_cs_survival(n = 530, p = 0.2, c = 0.85, n.predictors = 10, nsim = 100, parallel = TRUE)
}
\seealso{
samplesizedev_survival
}
