% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_performance.R
\name{expected_performance}
\alias{expected_performance}
\title{Calculate the expected calibration slope, MAPE and other performance metrics for a given  sample size}
\usage{
expected_performance(
  outcome = "Binary",
  n,
  phi,
  c,
  p,
  gamma = rep(1/p, p),
  nval = 25000,
  nsim = 1000,
  parallel = TRUE,
  method = "MLE",
  long = FALSE,
  approx = FALSE,
  x = NULL,
  y = NULL,
  threshold = NULL,
  individual_predicted_probability = NULL,
  x_individual = NULL
)
}
\arguments{
\item{outcome}{(character) The type of outcome (''Binary'' or ''Survival'')}

\item{n}{(numeric) The sample size}

\item{phi}{(numeric) The anticipated outcome prevalence (binary outcome) or proportion of events (survival outcome)}

\item{c}{(numeric) The anticipated c-statistic (binary outcome) or c-index (survival outcome)}

\item{p}{(numeric) The number of candidate predictor variables}

\item{gamma}{(numeric) Relative strength of predictors (default=rep(1/p,p); same length as p, must sums up to 1)}

\item{nval}{(numeric) Size of validation data (default=25000; use at least 10000)}

\item{nsim}{(numeric) The number of simulations (default=1000; use at least 500 to ensure small Monte Carlo simulation error)}

\item{parallel}{(logical) parallel processing to speed up computations (default=TRUE)}

\item{method}{(character) the fitting method (default="MLE"; currently the only option. Other fitting methods will be added in future versions)}

\item{long}{(logical) Extract results for all simulations instead of just averages}

\item{x}{(numeric) Matrix of Covariate values based on a user-defined DGM}

\item{y}{(numeric) Vector of responses based on a user-defined DGM}

\item{threshold}{(numeric) A threshold to calculate sensitivity (default=outcome prevalence)}

\item{individual_predicted_probability}{(numeric) An individual predicted probability of interest (default at median predicted probability) to obtain the uncertainty due to finite sample size (if x_individual_predicted_probability had not been provided)}

\item{x_individual_predicted_probability}{(numeric) a vector of covariate values for a individual of interest}
}
\value{
Data frame df with the following elements:

Input sample size (n)

Specified outcome prevalence (phi)

Specified C-statistic / C-index

Specified Number of predictors

Expected calibration slope (mean_CS)

Standard deviation of the CS (sd_CS)

Probability of obtaining a model with calibration slope between 0.85 and 1.15

Expected MAPE (MAPE)

Standard deviation of the expected MAPE (sd_MAPE)

Expected optimism in R square Nagelgerke (optimism_R2_Nag)

Expected AUC

Standard deviation of the expected AUC

Standard deviation of the average predicted risk

Expected Brier score

Sensitivity at the specified threshold (Sensitivity)

Net-Benefit at the specified threshold

Expected Individual predicted risk at the specified percentile

Standard deviation of the Individual predicted probability (IPP)
}
\description{
This function calculates the sampling distribution of calibration slope and MAPE and other metrics given key data and model characteristics
(outcome prevalence, c-statistic and number of predictors). It takes approximately 15 seconds to run (binary outcome).
}
\examples{
# expected_performance(outcome="Binary", n = 530, phi = 0.2, c = 0.85, p = 10, nsim = 100, parallel = FALSE)

# Prefer parallel computing with >2 cores that ensure faster running
# expected_performance(n = 530, phi = 0.2, c = 0.85, p = 10, nsim = 100, parallel = TRUE)
}
\seealso{
samplesizedev
}
