% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_mu_sigma2.R
\name{find_mu_sigma2}
\alias{find_mu_sigma2}
\title{Calculate the means and variance of a normal linear predictor given values of C and prevalence}
\usage{
find_mu_sigma2(
  target.prev,
  target.c,
  min.opt = c(-10, 0),
  max.opt = c(0.02, 5),
  tol = 1e-05,
  check = FALSE
)
}
\arguments{
\item{target.prev}{(numeric) The target prevalence}

\item{target.c}{(numeric) The target aUC}

\item{check}{(logical) check using a large dataset whether mean and variance correspond to the target values}
}
\value{
mean

variance

resulting C-statistic (if check = TRUE)

resulting prevalence (if check = TRUE)
}
\description{
This function calculates the mean and variance of a normal linear predictor that corresponds to the specified values of C and prevalence (binary outcome).
}
