% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalised_modified.R
\name{mod_penal_ave_foreach}
\alias{mod_penal_ave_foreach}
\title{Modified tuning
mod_penal_ave_foreach}
\usage{
mod_penal_ave_foreach(
  x,
  y,
  bn = 50,
  method = "ridge",
  f = nfolds/(nfolds - 1) - 1,
  parallel = TRUE,
  nfolds = 10,
  boot = TRUE
)
}
\arguments{
\item{x}{a matrix, the design matrix}

\item{y}{a vector, the binary response variable}

\item{bn}{a scalar, the number of bootstrap iterations}

\item{method}{a string, values "ridge" or "lasso"Numeric where to trim (proportion)}

\item{parallel}{a logical, TRUE for parallel computing}

\item{nfolds}{a scalar, the number of cross-validation folds}

\item{boot}{a logical, TRUE means sample with replacement (the default)

Outputs: a list with two elements.
- beta.boot is the calculated vector of regression coefficients
- lambda.boot is vector of beta's}
}
\description{
Modified tuning
mod_penal_ave_foreach
}
