% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesizedev.R
\name{samplesizedev}
\alias{samplesizedev}
\title{Sample Size Required to Develop a Risk Prediction model for Binary and Survival outcomes}
\usage{
samplesizedev(
  outcome = "Binary",
  S = 0.9,
  p,
  c,
  n.predictors,
  nval = 25000,
  nsim = 1000,
  parallel = TRUE
)
}
\arguments{
\item{outcome}{(character) The type of outcome (''Binary'' or ''Survival'')}

\item{S}{(numeric) The target expected calibration slope}

\item{p}{(numeric) The anticipated outcome prevalence (binary outcome) or proportion of events (survival outcome)}

\item{c}{(numeric) The anticipated c-statistic (binary outcome) or c-Index (survival outcome)}

\item{n.predictors}{(numeric) The number of candidate predictor variables}

\item{nval}{(numeric) Size of validation data (at least 10000)}

\item{nsim}{(numeric) The number of simulations (at least 500, default value 1000 to ensure small simulation error)}

\item{parallel}{(logical) parallel processing to speed up computations (default=TRUE)}
}
\value{
n: the required sample size
}
\description{
This function calculates the sample size required to achieve an expected Calibration Slope (S), given anticipated features of the data and the model. These
features are the outcome prevalence (p) and c-statistic (c) for a Binary outcome  or the proportion of events at a given time point (p) and c-index (c)  for a survival outcome,
and the number of candidate predictor variables (n.predictors).

The required value for the (average) calibration slope should be at least S=0.9. On the other hand, the input
values for p, c, and n.predictors are application-dependent.

The calculation takes approximately one minute for binary outcomes and 2-4 minutes for survival outcome (depending on the number of predictors).

We suggest that the variability in the calibration
slope, which is an indicator of model stability for the calculated sample size, be subsequently checked using the function 'expected_cs'. For binary outcomes, this
check also provides the expected Mean Absolute Prediction Error (MAPE) that corresponds to the calculated sample size.
}
\examples{
# Binary Outcome: Find the sample size required for an average calibration slope of S = 0.9
  samplesizedev(outcome="Binary", S = 0.9, p = 0.2, c = 0.85, n.predictors = 10,  nsim = 500, parallel = FALSE)

# Binary Outcome: Prefer parallel computing with >2 cores that ensure faster running
# samplesizedev(outcome="Binary", S = 0.9, p = 0.2, c = 0.85, n.predictors = 10,  nsim = 500, parallel = TRUE)

# Binary Outcome: Check the expected MAPE and Calibration Slope for the selected size
# expected_cs (outcome= "Binary", n = 530, p = 0.2, c = 0.85, n.predictors = 10, nsim = 500, parallel = TRUE)

# Survival Outcome: Find the sample size required for an average calibration slope of S = 0.9
# samplesizedev(outcome = "Survival", S = 0.9, p = 0.2, c = 0.85, n.predictors = 10,  nsim = 500, parallel = TRUE)

# Survival  Outcome: Check the expected MAPE and Calibration Slope for the selected size
# expected_cs(outcome = "Survival", n = 390, p = 0.2, c = 0.85, n.predictors = 10, nsim = 500, parallel = TRUE)


}
\seealso{
expected_cs
}
