% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampsizedev.R
\name{sampsizedev}
\alias{sampsizedev}
\title{Sample size required to develop a risk prediction model for binary outcomes}
\usage{
sampsizedev(
  S,
  c,
  p,
  n.predictors,
  nval = 25000,
  nsim = 1000,
  tol = 20,
  parallel = TRUE
)
}
\arguments{
\item{S}{(numeric) The target expected calibaration slope}

\item{c}{(numeric) The antcipated C-statistic}

\item{p}{(numeric) The anticipated outcome prevalence}

\item{n.predictors}{(numeric) The number of candidate predictor variables}

\item{nval}{(numeric) Size of validation data (at least 10000 )}

\item{nsim}{(numeric) Number of simulations (at least 500)}

\item{ncores}{(numeric) Number of processing computer cores}
}
\value{
n: the required sample size
}
\description{
This function calculates the sample size required to achieve an expected Calibration Slope (S), given anticipated features of the data and the model
(outcome prevalence, C-statistic and number of predictors).
}
\examples{
# Find the sample size
n <- sampsizedev(S=0.9, p = 0.2, c = 0.85, n.predictors = 10,  nsim = 100, parallel=FALSE)

# Fast run with parallel computing
# n <- sampsizedev(S=0.9, p = 0.2, c = 0.85, n.predictors = 10,  nsim = 100, parallel=TRUE)

# Check also the MAPE for the selected size
expected_cs_mape(n, p = 0.2, c = 0.85, n.predictors = 10, nsim = 100, parallel=FALSE)



}
\seealso{
expected_cs_mape()
}
