% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesizedev.R
\name{samplesizedev}
\alias{samplesizedev}
\title{Sample size required to develop a risk prediction model for binary outcomes}
\usage{
samplesizedev(
  S,
  c,
  p,
  n.predictors,
  nval = 25000,
  nsim = 1000,
  parallel = TRUE
)
}
\arguments{
\item{S}{(numeric) The target expected calibration slope}

\item{c}{(numeric) The antcipated C-statistic}

\item{p}{(numeric) The anticipated outcome prevalence}

\item{n.predictors}{(numeric) The number of candidate predictor variables}

\item{nval}{(numeric) Size of validation data (at least 10000 )}

\item{nsim}{(numeric) The number of simulations (at least 500, default value 1000 to ensure small simulation error)}

\item{parallel}{(logical) parallel processing to speed up computations (default=TRUE)}
}
\value{
n: the required sample size
}
\description{
This function calculates the sample size required to achieve an expected Calibration Slope (S), given anticipated features of the data and the model
(outcome prevalence, C-statistic and number of predictors). It takes approximately one minute to run. Ideally it should be followed by checking also
the Mean Absolute Prediction Error that corresponds to the calculated sample size.
}
\examples{
# Find the sample size
n <- samplesizedev(S=0.9, p = 0.2, c = 0.85, n.predictors = 10,  nsim = 100, parallel=FALSE)

# Fast run with parallel computing
# n <- samplesizedev(S=0.9, p = 0.2, c = 0.85, n.predictors = 10,  nsim = 100, parallel=TRUE)

# Check the expected MAPE and Calibration Slope for the selected size
# expected_cs_mape(n, p = 0.2, c = 0.85, n.predictors = 10, nsim = 100, parallel=FALSE)



}
\seealso{
expected_cs_mape()
}
