% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesizedev.R
\name{samplesizedev}
\alias{samplesizedev}
\title{Sample Size Required to Develop a Risk Prediction model for Binary Outcomes}
\usage{
samplesizedev(
  outcome = "Binary",
  S = NULL,
  MAPE = NULL,
  phi,
  c,
  p,
  gamma = rep(1/p, p),
  nval = 25000,
  nsim = 1000,
  parallel = TRUE
)
}
\arguments{
\item{outcome}{(character) The type of outcome (''Binary'' ; ''Survival'' to be added in later versions)}

\item{S}{(numeric) The target expected calibration slope}

\item{phi}{(numeric) The anticipated outcome prevalence of the binary outcome (or proportion of events for survival outcome)}

\item{c}{(numeric) The anticipated c-statistic for binary outcome (or c-index for survival outcome)}

\item{p}{(numeric) The number of candidate predictor variables}

\item{gamma}{(numeric) The Relative strength of predictors (default=rep(1/p,p); same length as p, must sum up to 1)}

\item{nval}{(numeric) Size of validation data (default=25000; use at least 10000)}

\item{nsim}{(numeric) The number of simulations (default=1000; use at least 500 to ensure small simulation error)}

\item{parallel}{(logical) parallel processing to speed up computations (default=TRUE)}
}
\value{
n$sim: the required sample size by simulation

n$rvs: the required sample size by the RvS formulae (only for comparison)
}
\description{
This function calculates the sample size required to achieve an expected Calibration Slope (S), given anticipated features of the data and the model. For a binary outcome these
features are the outcome prevalence (phi) and c-statistic (c) and the number of candidate predictor variables (p).
Future versions will  accept the anticipated proportion of events at a given time point (phi) and c-index (c)  for a time to event outcome.

The required value for the (average) calibration slope should be at least S=0.9. On the other hand, the input
values for phi, c, and p are study-dependent.

The calculation usually takes  a minute or less for a binary outcome.

The variability in the calibration slope, can be seen as an indicator of model stability
for the calculated sample size. This can be evaluated using the function 'expected_cs'. For Binary
outcomes, this check also provides the expected Mean Absolute Prediction Error (MAPE) that corresponds
to the calculated sample size.
}
\examples{

# Binary Outcome: Prefer parallel computing  that ensures faster run
# samplesizedev(outcome = "Binary", S = 0.9, phi = 0.2, c = 0.85, p = 10, parallel=TRUE)

# Binary Outcome: Check the expected MAPE and Calibration Slope for the selected size
# expected_cs (outcome = "Binary", n = 530, phi= 0.2, c = 0.85, p = 10)

}
\seealso{
expected_cs
}
